backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
google-genai